requirements.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==7.1.2
Flask==1.1.4
Flask-Login==0.4.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.4
MarkupSafe==1.1.1
psycopg2==2.9.9
requests==2.31.0
SQLAlchemy==1.4.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==0.16.1
WTForms==3.0.1
zipp==3.17.0
